feat: logdialogeditor
